Revert "Revert "✨ [FEAT] 선배 개인페이지 API 수정""
