+++ NadoSunbae-iOS/NadoSunbae-iOS/Screen/Mypage/MypageMain/VC/MypageMainVC+TV.swift
@@ -15,7 +15,8 @@ extension MypageMainVC: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: BaseQuestionTVC.className, for: indexPath) as? BaseQuestionTVC else { return UITableViewCell() }
-        cell.setData(data: self.questionList[indexPath.row])
+        let questionData = self.questionList[indexPath.row]
+        cell.setPostData(data: PostListResModel(postID: questionData.postID, type: nil, title: questionData.title, content: questionData.content, createdAt: questionData.createdAt, majorName: "", writer: questionData.writer, isAuthorized: false, commentCount: questionData.commentCount, like: questionData.like))
         cell.layoutIfNeeded()
         
         return cell
@@ -35,6 +36,7 @@ extension MypageMainVC: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         self.navigator?.instantiateVC(destinationViewControllerType: DefaultQuestionChatVC.self, useStoryboard: true, storyboardName: Identifiers.QuestionChatSB, naviType: .push) { questionDetailVC in
             questionDetailVC.hidesBottomBarWhenPushed = true
+            questionDetailVC.isAuthorized = true
             questionDetailVC.naviStyle = .push
             questionDetailVC.postID = self.questionList[indexPath.row].postID
         }
