+++ NadoSunbae-iOS/NadoSunbae-iOS/Screen/Mypage/MypageUser/VC/MypageUserVC+TV.swift
@@ -15,7 +15,8 @@ extension MypageUserVC: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = BaseQuestionTVC()
-        cell.setData(data: questionList[indexPath.row])
+        let questionData = self.questionList[indexPath.row]
+        cell.setPostData(data: PostListResModel(postID: questionData.postID, type: nil, title: questionData.title, content: questionData.content, createdAt: questionData.createdAt, majorName: "", writer: questionData.writer, isAuthorized: false, commentCount: questionData.commentCount, like: questionData.like))
         cell.layoutIfNeeded()
 
         return cell
@@ -40,6 +41,7 @@ extension MypageUserVC: UITableViewDelegate {
                 questionDetailVC.hidesBottomBarWhenPushed = true
                 questionDetailVC.naviStyle = .push
                 questionDetailVC.postID = self.questionList[indexPath.row].postID
+                questionDetailVC.isAuthorized = self.userInfo.userID == UserDefaults.standard.integer(forKey: UserDefaults.Keys.UserID)
             }
         }
     }
